[tests] E: Reuse CapabilityGuard Implementation
